Move log outside ophyd_async/core
